requirements.txt
Django==1.4
PIL==1.1.7
South==0.7.4
boto==2.3.0
django-storages==1.1.4
docutils==0.8.1
httplib2==0.7.4
oauth2==1.5.211
psycopg2==2.4.5
sorl-thumbnail==11.12
wsgiref==0.1.2
